test(coffee): share repository contract suite
- move the repository contract helper into coffee-core application testing
- point in-memory, SQLite, and Drizzle Postgres suites at the same contract
- leave Postgres-specific migration and id-generator checks local to its adapter
