+++ apps/backend/test/contracts/in-memory-repositories.test.ts
@@ -1,6 +1,6 @@
 import * as Effect from "effect/Effect";
 import { InMemoryCoffeeRepositoriesLive } from "@effect-coffee-shop/coffee-external-in-memory";
-import { defineRepositoryContract } from "./repository-contract.ts";
+import { defineRepositoryContract } from "@effect-coffee-shop/coffee-core/application/testing/repository-contract";
 
 defineRepositoryContract("in-memory repositories", (effect) =>
   Effect.runPromise(effect.pipe(Effect.provide(InMemoryCoffeeRepositoriesLive))),
